#547 - Split dashboard into user and admin areas
- Merge dashboard module into core ui module to enable tigher coupling with menu bar
