+++ inception-ui-core/src/main/java/de/tudarmstadt/ukp/inception/ui/core/bootstrap/BootstrapWizard.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.app.bootstrap;
+package de.tudarmstadt.ukp.inception.ui.core.bootstrap;
 
 import org.apache.wicket.Component;
 import org.apache.wicket.extensions.wizard.IWizardModel;
