+++ inception-ui-core/src/main/java/de/tudarmstadt/ukp/inception/ui/core/dashboard/AdminDashboardPageMenuItem.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.ui.dashboard;
+package de.tudarmstadt.ukp.inception.ui.core.dashboard;
 
 import static java.util.Arrays.asList;
