+++ inception-ui-core/src/main/java/de/tudarmstadt/ukp/inception/ui/core/dashboard/DashboardMenu.html
