--- inception-ui-dashboard/marker-wicket-module
@@ -1,1 +1,0 @@
-Marker file which activates the profile "wicket-module" from the parent POM.
